Fix the most recent round of jQuery validator errors
The malformed HTML caused the bots and user API key forms to be
merged.  When the user clicked the "Show your API key" button, a
form-related event gets triggered, which jquery.validate tries to
handle.  When it does so, it ends up throwing an exception because
the form element involved in the event does not have an associated
validator object attached to it.

(imported from commit 26100a443603dc8a6cc23b9f3e825632149faf7f)