[third] Patch bootstrap-typeahead to hide when the source fn returns false.
On the next line, it checks and returns immediately if the source returns
false, but it should first hide the existing box or else it's just left
there until the next typeahead refresh.

This fixes the issue with the typeahead not closing properly when a space
is typed that was introduced in 7aa3f762.

(imported from commit cbe77e2152d889fe53d874e97f41cfc725166929)