Fix collapse issue introduced by ca50035 while keeping the effect of its fix.
That commit intended to fix a bug where it jumps to the wrong scroll position
(offset by the difference in height from uncollapsed to collapsed) if you load
collapsed messages at the top, either by scrolling up or when it automatically
fetches more 10s after page load, because before, it updated the scroll
position before processing collapsing, which changes the height of messages.

Updating collapsing needs to happen before the viewport is adjusted, but also
needs to happen after the new messages get appended/prepended to the document
because they need to be attached to the document for getClientBoundingRect to
work.

(imported from commit 44d6a6135524e658084ffcf7d880ba898b577e6d)