Speed up subscription table rendering
Previously, we'd render all the rows of the table except the first
(which is for creating new subscriptions) and append them all to the
table.  This ends up being really slow.  We now instead render the
entire tbody element and replace the existing tbody element.

When profiling on my laptop, this reduces the rendering time for a
few thousand streams from ~62 seconds to ~2 seconds.

(imported from commit 83a48e0e1f776f7663343662157298e89165ece3)